Fix #8: Bildeinfügen in den Produktbeschreibungen funktioniert nicht
The problem is that the we call `stripslashes()` and `addslashes()`
somewhat arbitrarily. Assuming that we require PHP ≥ 5.4.0 (or at least
that all `magic_quotes_*` settings are off) we actually don't need
stripslashes()` and `addslashes()` at all; instead, in some cases we
use `addcslashes()` where necessary. We keep a single occurrence of
`addslashes()`, though, namely where we print the product name as
argument to a JS function call in an inline event handler attribute. We
should really not do this at all, and it's probably best not to use any
inline event handlers altogether, but we can fix that later.