Improve ordering of products
Changing the order of products in the back-end is horrible. To really
fix this, we would need to completely rework the products table, so
that it is a single form. Since that appears too much a change for
version 1, we deploy a workaround, namely that the swap requests are
now sent via XHR (aka. Ajax), if the browser is modern, which is faster
and avoids scrolling to the top after each swap operation.