Close #13: Move help to standalone HTML files
We move the help that is integrated into the plugin backend to
customary standalone HTML files to ease translation and to allow for
individual styling without cluttering the plugin stylesheet.

Since the changelog ("What's new") doesn't necessarily belong to the
help files (aka. user manuals), we move it to a simple plain text
CHANGELOG.