Close #64: Move payment module settings and lang to plugin level
This simplifies the payment modules, as there is no need for
configuration forms any more, and we're close to be able to flatten the
folder structure of the payment modules.

While we're at it, we move the currency code option beneath the default
currency option (because these have to match), and we change the
environment option to a boolean sandbox option to avoid complex
translations of the enumeration values.