Fix #35: Insufficient date localization
In an ideal world we'd use `DateFormatter` and be done, but in this
world ext/intl is often unavailable, apparently. We could still use it
and add a fallback solution, but as we need the date only for the
preliminary bills, we simply use `date()` with a format specified in
the language configuration, and call it a day.