Fix #93: bill.rtf can't deal with Unicode except for ISO-8859-1
To convert UTF-8 to ISO-8859-1 (via `utf8_decode`) is insufficient, so
we convert all non ASCII characters to RTF Unicode escape sequences.
While we have to program the RTF escaping algorithm ourselves, we reuse
`mb_convert_encoding()` to convert to UTF-16 first, therefore requiring
ext/mbstring now, what shouldn't be a problem as it's required by
CMSimple_XH 1.7.0 anyway.

We also remove the special handling of currency symbols, because it is
already catered to as of this commit.