Close #96: Improve customizability of template.rtf
Instead of having a %rows% parameter in `template.rtf` and replacing
it with hard-coded rows, we support to have a *single* row template for
the order lines in `template.rtf`, which is automatically duplicated
and filled in. We use short lower case placeholders to eschew breaking
the layout of the template itself.