Solve the mistery
The TODO obviously refers to a live server with PHP 4 and a local XAMPP
installation with PHP 5. Under PHP 4 objects were passed by value, but
as of PHP 5.0.0 they're passed by identity, what makes all the
difference here. As we're not supporting PHP 4 anymore, we can simplify
the code to its original intention.