Replace bill.rtf with bill.csv
RTF has several issues so we replace it with CSV. Actually, we do not
necessarily intend to replace RTF altogether, but rather consider to
offer both alternatives. Until a decision in this regard has been made,
this commit serves as POC for the CSV solution.

There is `templates/frontend/confirmation_email/billtemplate.csv` which
can be used as is or customized by the webmaster. Unnecessary
placeholders can simply be removed, and everything can be reordered to
fit individual needs. The only specialty is the order line row, which
always has to be a single row, and must not contain other placeholders
then the default ones. However, reordering and/or removing these
placeholders is allowed.

We're currently hard-coding the separator character as `;`.

Whether the generated `bill.csv` has a BOM depends on whether the
template has one.