Replace custom back-end menu with native plugin menus
Instead of the manually created menu we use the plugin menus of
CMSimple_XH. While the integrated plugin menu is always available, the
classic plugin menu is only shown if the plugin administration is
explicitly requested, but not on the shop page, because we save space
and it looks nicer. It is also noteworthy that the plugin
administration defaults to the about page while the shop page defaults
to the product overview.