Keep JavaScript out of HTML
For maintainability and resilience we move the JavaScript out of the
HTML. Since we require a bit of JavaScript in the front-end we deliver
the JavaScript always what shouldn't be an issue. We rename
`xhsbackend.min.js` accordingly.

It is noteworthy that all JavaScript is restricted to modern browsers
now, for instance, IE 9 is not supported anymore.