Add optional differing delivery address
We allow users to enter a differing delivery address in the contact
data. All fields are optional and the delivery address is accepted, if
all fields are filled out, in which case it is displayed in the final
confirmation step and also in the e-mails. The respective values are
also available in the bills, but we have not catered to that in the
billtemplates.