Overhaul checkout navigation
We implement safe requests (such as showing the cart and proceeding to
the customer data) as GET requests. We also deploy the PRG pattern for
the other requests, so that users can use the browser's forward and
back buttons instead of the ones displayed at the bottom of the forms.
To ease handling of these requests without using `$_REQUEST` we move
the `xhsCheckout` parameter into the URL.

We also make sure that only allowed checkout steps can be requested;
otherwise we're redirecting to the appropriate step.