Close #28: Link preview image to product details
We link the preview image to the product details in the front-end
product list, instead of the products image.

We also take the opportunity to unify the handling of linked images by
introducing `View::linkedImage()` which allows us to get rid of
`Product::getImage()` and `Product::getPreviewPicture()` which do not
belong to this model class. We also dump `Product::getBestPicture()`.