Country specific forwarding expenses on multilang sites
If a buyer has already completed step 2 of the checkout, but changes to
another language before actually ordering, the country usually becomes
invalid (because the name has changed, e.g. "Germany" → "Deutschland").
This isn't a real problem, but it potentially causes erroneous behavior
of `splitForwardingExpenses()`, which we make more robust.