Calculate taxes as late as possible
Formerly the taxes have basically been calculated for each single item
in the cart what easily leads to rounding errors. Therefore we sum up
the value of goods (according to the VAT rate) first, and calculate the
taxes on this sum. Apparently, the new calculation is more exact, and
so we have to fix the respective tests.