Introduce `Decimal` class
For now the `Decimal` class' primary responsibility is to normalize
numeric values to a BC Math conforming format. This could, of course,
also be accomplished with a single function, but using a class instead
gives us some type safety, and offers the possibility to abstract over
BC Math if so desirable.