Fix #132: Inconsistent representation of VAT rates
The inconsistency that sometimes the VAT rates where shown with
exactly two decimal places has already been fixed in 2463f38. We
internationalize the percentage values with this commit. The chosen
solution is slightly fragile, as it relies on string values so we have
quickly thrown in an `assert()` which we have to fix ASAP (see #134).