Close #134: Cater to assert()s before shipping
Since there is no portable way to write `assert`ions for PHP 5 and 7[1]
we use `trigger_error()` to signal programming logic errors. We do not
abort program execution, though, but rather raise `E_USER_WARNING` to
better simulate the effect of disabled assertion checking on production
systems.

We should migrate to proper assertions as soon as we require PHP 7.

[1] While PHP 5 requires assertion expressions to be written as
strings, this "feature" is deprecated as of PHP 7.2.