Add boolean %MAY_FORWARD_EMAIL%
In Germany and perhaps other countries, it appears to be a legal
requirement to obtain the buyer's approval to forward their email
address to the parcel service. Therefore we add a respective checkbox
to the customer data form, and forward the information to the "bill".

See https://github.com/cmsimple-xh/xhshop/issues/15#issuecomment-318270067.