Fix Paypal button URLs to the shop
Besides that we cannot rely on the `HTTP_REFERER` anyway, we use more
detailed URLs to take the customers to the appropriate place. If they
cancel, they probably do not want to pay via Paypal for some reason, so
we take them back to the customers data where they can choose another
payment method. If they have payed and go back to the shop we always
take them to the thank you page, regardless of whether the appropriate
IPN has been received or not. Otherwise a delayed IPN would still show
their items in the cart, and they might see the thank you page on their
next visit to the shop which could be days into the future.

In the somewhat unlikely case that they don't return to the shop after
payment, their cart and customer data won't be purged, but that appears
to be the lesser evil.