Make IPN handshake more resilient
Regardless of whether the handshake succeeded or not, we terminate the
script right away, clean all output buffers and respond with an
appropriate status code, so that Paypal is informed about whether to
resend the IPN or not.

We also log any unexpected handshake failures to CMSimple_XH's log
file, which might be helpful for debugging purposes.