`finishCheckOut()` needs special treatment when triggered by IPN
In this case we must not "relocate", i.e. send relocation headers,
since Paypal would send the IPN again, but we couldn't use it anyway.
And more importantly, we must not check for the CSRF token, since this
is not send from Paypal, because it is not necessary – proper
authorization is already guaranteed by the IPN handshake.