Fix #129: Shop crashes if tmp_orders/ is missing
During checkout, if Paypal is chosen as payment method and the folder
`tmp_orders/` is missing, the shop `die()`s. Of course, this is a
no-go. Instead of triggering an error message which wouldn't be helpful
for the buyer anyway, we try to create the folder and go on.

We do not care about the rare cases where the folder doesn't exists and
cannot be created – after all, there is already a system check for the
existence and writability of this folder.