Fix #148: Invalidate carts if config is saved
To avoid issues with changes to the shop configuration during ordering,
we invalidate all shopping carts as soon as the configuration is saved,
what happens whenever the shop's publishing state changes.

The shop providers have to regard, that for any non trivial change to
the language specific configuration or to the products, the shop *has*
*to* be switched to maintenance mode. Trivial changes would be, for
instance, fixing a typo in a product description, or adding a new
product.