Tighten handling of decimals
We replace `::__toString()` with `::toString()` to eschew unintended
casts which might result in inexact arithmetic. We also do not allow to
pass floats to the constructor anymore, to avoid floating point issues.
The only place where floats actually might occur are in `catalog.php`
from old Wellrad, so we add proper float to decimal conversion for this
case.

Additionally, we extend the accepted string format for the constructor,
but raise a warning otherwise in the back-end; in the front-end we
throw an uncaught exception – it's better to sell nothing, than to
allow erroneous orders to be made. The shop provider is responsible to
verify that the system check is green, and that the product information
is appropriate. If that is so, there Exceptions are not supposed to be
thrown.