Add EmlBillWriter
The `EmlBillWriter` allows to add a prepared email as attachment to the
notification email, which can later be send to the customer (for
instance, as delivery notice).

The first line of `billtemplate.eml` contains the subject of the email,
and the rest is the HTML which will constitute the email body.