Allow multiple "bill" attachments
Instead of restricting users to a single "bill" attachment to the
notification email, we allow for an arbitrary amount in whatever format
that is supported. The shop provider has to enter the desired "bills"
in the language configuration as comma separated list, and has to
provide respective templates in `templates/confirmation_email/`. The
actual templates have to have an additional `.tpl` between the basename
and the extension to avoid confusion regarding the filenames we're
talking about (documentation, support, etc.)

This change has the additional benefit of letting the shop provider use
more appropriate filenames than "bill", and also allows for proper
localization.