Fix add to cart behavior
Adding goods to the cart should actually *add* them instead of
*replacing* them. For instance, adding a certain item twice, should
result in having two of these items in the cart. On the other hand,
adding a negative amount should remove this amount from the items in
the cart.

During the first step of the final checkout, the customer still can
change the amounts as before, i.e. the form works by replacing the
items here.