Fix #217: variant bug when page has several cartButtons
We have to properly reset the view params, so that we do not
accidentially pick up params from an earlier rendering.  Unfortunately,
we are using some dynamic properties, what makes that slightly harder
and more brittle than necessary.

We also have to make sure that there is only ever a single element with
a certain ID.  Since there is no need to have an ID for the cart
buttons, we change that to a class.  This is a BC break, though, due to
customizations.