Fix #218: If Allow Show All is off, still the whole Shop is shown
If `allow_show_all` is disabled, we change the default category to the
first available category on the fly, if the chosen default category is
not contained in the available categories.  We also make sure that we
do not show the "show all" category for the default category selection
in the category editor.