Update further development tools
We need to be able to run all these tools under PHP 8.0 at least, so we
bite the bullet and update.  This requires to update to PHPUnit 9.5
which requires us to change coverage.xml.  We also upgrade PHPCS to 3,
what requires us to call the command line script directly (due to a
Phing limitation).  And unfortunately, we have to drop PHPCompatInfo;
the latest version I've ever managed to use on Windows is 5.4, and that
would be okay, but it does apparently only support a single folder to
scan, what is too limited for our purposes (or would also scan vendor/).