Fix #200 (Don't use <noscript>)
The category select box in the back-end is usually powered by
JavaScript, but is supposed to work without it.  However, if the
respective script can't be loaded, the button is not shown, since it
is hidden in a `<noscript>` element, so the select box is
dysfunctional.  We fix this by removing the `<noscript>`, and instead
hide the button via JavaScript.